Conflicts:
	Assets/Materials/Swatches/Material Color Layer (Additive).mat
	Assets/Materials/Swatches/Material Color Layer (Mulitplicative).mat
	Assets/Materials/Swatches/Primary.mat
	Assets/Materials/Swatches/Primary_Dark.mat
	Assets/Materials/Swatches/Primary_Med.mat
	Assets/Materials/Swatches/Secondary.mat
	Assets/Materials/Swatches/Secondary_Dark.mat
	Assets/Materials/Swatches/Secondary_Med.mat
	Assets/Materials/Swatches/Tri.mat
	Assets/Materials/Swatches/Tri_Dark.mat
	Assets/Materials/Swatches/Tri_Med.mat
	Assets/_Scene/BetaLevels 1/LV3.unity
	Library/CurrentLayout.dwlt
	Library/assetDatabase3
	Library/expandedItems
	Library/guidmapper
	Library/shadercompiler-1.log

+++ Assets/Scripts/CameraController.cs
@@ -6,7 +6,7 @@ public class CameraController : MonoBehaviour {
 	public GameObject player;
 	public Vector3 offset = new Vector3(0, 1.5f, -10);
 
-    public int minX, minY, maxX, maxY;
+    public float minX, minY, maxX, maxY;
 
 	// Use this for initialization
 	void Awake () {
